Remove dependency on fordefi
