+++ automated-orders/package.json
@@ -13,9 +13,9 @@
     "build": "yarn lint && tsc --build --force --verbose",
     "lint": "eslint --ext .ts --ignore-path .eslintignore .",
     "build:blockchain-lib": "yarn workspace @warden-automated-orders/blockchain build",
-    "build:fordefi-lib": "yarn workspace @warden/fordefi-library build",
+    "build:aws-kms-signer": "yarn workspace @warden/aws-kms-signer build",
     "build:utils": "yarn workspace @warden-automated-orders/utils build",
-    "build:libs": "yarn build:utils && yarn build:blockchain-lib && yarn build:fordefi-lib",
+    "build:libs": "yarn build:utils && yarn build:blockchain-lib && yarn build:aws-kms-signer",
     "build:scheduler": "yarn workspace @warden-automated-orders/scheduler build",
     "build:relayer": "yarn workspace @warden-automated-orders/relayer build",
     "scheduler": "yarn workspace @warden-automated-orders/scheduler start",
