+++ automated-orders/packages/aws-kms-signer/src/types/configuration.ts
@@ -1,0 +1,4 @@
+export interface IAwsKmsConfiguration {
+  awsKmsRegion: string;
+  awsKmsKeyId: string;
+}
