+++ automated-orders/packages/aws-kms-signer/tsconfig.json
