--- automated-orders/packages/fordefi-library/src/types/blackbox-signature/signature.signerType.ts
@@ -1,1 +1,0 @@
-export type SignerType = 'initiator' | 'api_signer' | 'end_user';
