--- automated-orders/packages/fordefi-library/src/types/user.ts
@@ -1,9 +1,0 @@
-export interface User {
-  id: string;
-  user_type: string;
-  name: string;
-  state: string;
-  role?: string;
-  external_id?: string;
-  email?: string;
-}
