--- automated-orders/packages/fordefi-library/src/types/vault/configuration.ts
@@ -1,6 +1,0 @@
-export interface IFordefiConfiguration {
-  fordefiAPIEndpoint: string;
-  accessToken: string;
-  awsKmsRegion: string;
-  awsKmsKeyId: string;
-}
