--- automated-orders/packages/fordefi-library/src/types/vault/keyset.ts
@@ -1,8 +1,0 @@
-export interface Keyset {
-  /** The unique identifier of the keyset in the Fordefi platform. */
-  id: string;
-  /** The name of the keyset. */
-  name: string;
-  /** Represents a keyset scope in the Fordefi platform */
-  scope: string;
-}
