--- automated-orders/packages/fordefi-library/src/types/vault/vault.bitcoinChain.ts
@@ -1,1 +1,0 @@
-export type BitcoinChain = 'bitcoin_mainnet' | 'bitcoin_testnet';
