--- automated-orders/packages/fordefi-library/src/types/vault/vault.derivationInfo.ts
@@ -1,8 +1,0 @@
-import { MasterPublicKey } from './masterPublicKey.js';
-
-export interface VaultDerivationInfo {
-  /** The BIP 44 derivation path of the vault. */
-  derivation_path: string;
-  /** The public key of the vault. */
-  master_public_key: MasterPublicKey;
-}
