--- automated-orders/packages/fordefi-library/src/types/vault/vault.keyType.ts
@@ -1,1 +1,0 @@
-export type KeyType = 'ecdsa_stark' | 'ecdsa_secp256k1';
