--- automated-orders/packages/fordefi-library/src/types/vault/vault.type.ts
@@ -1,1 +1,0 @@
-export type VaultType = 'black_box' | 'cosmos' | 'evm' | 'solana' | 'utxo';
