--- automated-orders/packages/fordefi-library/src/types/vault/vaultRef.ts
@@ -1,9 +1,0 @@
-import { User } from '../user.js';
-
-export interface VaultRef {
-  id: string;
-  name: string;
-  address?: string;
-  state: string;
-  end_user?: User;
-}
